+++ lib/ansible/ruby/dsl_builders/block_spec.rb
@@ -84,7 +84,7 @@ describe Ansible::Ruby::DslBuilders::Block do
 
     subject { -> { evaluate } }
 
-    it { is_expected.to raise_error "Invalid method/local variable `foobar'. Only valid options are [:task, :become, :become_user, :ansible_when, :ignore_errors, :jinja] at line 1!" }
+    it { is_expected.to raise_error "Invalid method/local variable `foobar'. Only valid options are [:task, :become, :become_user, :ansible_when, :ignore_errors, :jinja]" }
   end
 
   context 'multiple register' do