+++ lib/ansible/ruby/dsl_builders/file_level.rb
@@ -34,6 +34,23 @@ module Ansible
           @tasks_builder.handler name, &block
         end
 
+        def _handled_eval(ruby_filename)
+          ruby_code = File.read ruby_filename
+          instance_eval ruby_code, ruby_filename
+          # error code
+          nil
+        rescue StandardError => error
+          only_user_code = error.backtrace_locations
+                                .select { |trace| trace.absolute_path == ruby_filename }
+                                .map { |trace| format_trace_line(trace) }
+          message = "#{error.message}\n****Error Location:****\n#{only_user_code.join("\n")}"
+          Exception.new message
+        end
+
+        def format_trace_line(trace)
+          "#{trace.path}:#{trace.lineno}"
+        end
+
         # any order/lazy result
         # :reek:NilCheck - when nil is the simplest way to check this
         def _result