+++ lib/ansible/ruby/dsl_builders/module_call_spec.rb
@@ -157,7 +157,7 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
 
       subject { -> { evaluate } }
 
-      it { is_expected.to raise_error "Can't use arguments [\"howdy\"] on this type of module at line 1!" }
+      it { is_expected.to raise_error "Can't use arguments [\"howdy\"] on this type of module" }
     end
 
     context 'no block' do
@@ -169,10 +169,10 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
 
       subject { -> { evaluate } }
 
-      it { is_expected.to raise_error 'You must supply a block when using this type of module at line 1!' }
+      it { is_expected.to raise_error 'You must supply a block when using this type of module' }
     end
 
-    context 'valid' do
+    context 'valid module' do
       before do
         klass_free_form = Class.new(Ansible::Ruby::Models::Base) do
           attribute :free_form
@@ -187,7 +187,7 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
         end
       end
 
-      context 'valid' do
+      context 'block, no jinja' do
         let(:ruby) do
           <<-RUBY
           command 'ls /stuff' do
@@ -235,7 +235,7 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
 
         subject { -> { evaluate } }
 
-        it { is_expected.to raise_error 'Expected only 1 argument for this type of module at line 1!' }
+        it { is_expected.to raise_error 'Expected only 1 argument for this type of module' }
       end
 
       context 'free form not supplied' do
@@ -249,7 +249,27 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
 
         subject { -> { evaluate } }
 
-        it { is_expected.to raise_error 'Expected 1 argument for this type of module at line 1!' }
+        it { is_expected.to raise_error 'Expected 1 argument for this type of module' }
+      end
+
+      context 'argument not found' do
+        let(:ruby) do
+          <<-RUBY
+          command 'ls /stuff' do
+            bar '/file1.conf'
+          end
+          RUBY
+        end
+
+        subject { -> { evaluate } }
+
+        it do
+          is_expected.to raise_error <<ERROR
+Unknown attribute 'bar' for Ansible::Ruby::Modules::Command.
+
+Valid attributes are: [:foo]
+ERROR
+        end
       end
     end
   end
@@ -270,7 +290,7 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
 
     subject { -> { evaluate } }
 
-    it { is_expected.to raise_error "Validation failed: Dest can't be blank at line 6!" }
+    it { is_expected.to raise_error "Validation failed: Dest can't be blank" }
   end
 
   context 'not found' do
@@ -286,6 +306,6 @@ describe Ansible::Ruby::DslBuilders::ModuleCall do
 
     subject { -> { evaluate } }
 
-    it { is_expected.to raise_error 'Unknown module foo_copy at line 2!' }
+    it { is_expected.to raise_error 'Unknown module foo_copy' }
   end
 end