+++ lib/ansible/ruby/dsl_builders/task.rb
@@ -69,18 +69,24 @@ module Ansible
         private
 
         def _process_method(id, *args, &block)
-          if id == :ansible_include
-            if @context == Models::Handler
-              raise "Can't call inclusion inside a handler(yet), only in plays/handlers"
-            else
-              raise "Can't call inclusion inside a task, only in plays/handlers"
-            end
-          end
-          # only 1 module, so don't try and do this again
-          no_method_error id, "Only valid options are #{_valid_attributes}" if @module
+          _check_context if id == :ansible_include
           mcb = ModuleCall.new
+          if @module && mcb.respond_to?(id)
+            # only 1 module allowed per task, give a good error message
+            raise "Invalid module call `#{id}' since `#{@module.ansible_name}' module has already been used in this task. Only valid options are #{_valid_attributes}"
+          elsif @module
+            no_method_error id, "Only valid options are #{_valid_attributes}"
+          end
           mcb.send(id, *args, &block)
           @module = mcb._result
+        end
+
+        def _check_context
+          if @context == Models::Handler
+            raise "Can't call inclusion inside a handler(yet), only in plays/handlers"
+          else
+            raise "Can't call inclusion inside a task, only in plays/handlers"
+          end
         end
 
         def method_missing_return(_id, _result, *_args)