+++ spec/ansible_lint/role_test.rb
@@ -1,7 +1,0 @@
-play 'the play name' do
-  hosts %w(host1 host2)
-
-  roles 'role1'
-
-  user 'centos'
-end