+++ spec/rake/no_nested_tasks/playbook_error.rb
@@ -1,16 +1,14 @@
 play 'the play name' do
   hosts %w(host1 host2)
 
+  ansible_iinclude 'inclusion.yml' do
+    variables howdy: 123
+  end
+
   task 'Copy something over' do
-    result = copy do
+    copy do
       src '/file1.conf'
       dest '/file2.conf'
     end
-
-    become
-    notify 'handler1'
-    changed_when "'no upgrade' in #{result.stdout}"
   end
-
-  user 'centos'
 end